Move the calls test doubles into `pimm.tests.testing`
`Passive` and `FakeCall` were hand-rolled per file; `FakeCall` is typed by the
call it stands in for rather than fixed at the grippers' `float`.
